More refactoring; cleaning up code, preparing for tests
