+++ config/letsencrypt/storage.go
@@ -86,43 +86,3 @@ func emailUsername(email string) string {
 	}
 	return email[:at]
 }
-
-/*
-// StorageDir is the full path to the folder where this Let's
-// Encrypt client will set up camp. In other words, where it
-// stores user account information, keys, and certificates.
-// All files will be contained in a 'letsencrypt' folder
-// within StorageDir.
-//
-// Changing this after the program has accessed this folder
-// will result in undefined behavior.
-var StorageDir = "."
-
-// Values related to persisting things on the file system
-const (
-	// ContainerDir is the name of the folder within StorageDir
-	// in which files or folders are placed.
-	ContainerDir = "letsencrypt"
-
-	// File that contains information about the user's LE account
-	UserRegistrationFile = "registration.json"
-)
-
-// BaseDir returns the full path to the base directory in which
-// files or folders may be placed, e.g. "<StorageDir>/letsencrypt".
-func BaseDir() string {
-	return filepath.Join(StorageDir, ContainerDir)
-}
-
-// AccountsDir returns the full path to the directory where account
-// information is stored for LE users.
-func AccountsDir() string {
-	return filepath.Join(BaseDir(), "users")
-}
-
-// AccountsDir gets the full path to the directory for a certain
-// user with the email address email.
-func AccountDir(email string) string {
-	return filepath.Join(AccountsDir(), email)
-}
-*/
