+++ config/letsencrypt/user.go
@@ -1,13 +1,17 @@
 package letsencrypt
 
 import (
+	"bufio"
 	"crypto/rand"
 	"crypto/rsa"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"io/ioutil"
 	"os"
+	"strings"
 
+	"github.com/mholt/caddy/server"
 	"github.com/xenolf/lego/acme"
 )
 
@@ -29,6 +33,7 @@ func (u User) GetPrivateKey() *rsa.PrivateKey {
 }
 
 // getUser loads the user with the given email from disk.
+// If the user does not exist, it will create a new one.
 func getUser(email string) (User, error) {
 	var user User
 
@@ -59,7 +64,7 @@ func getUser(email string) (User, error) {
 }
 
 // saveUser persists a user's key and account registration
-// to the file system.
+// to the file system. It does NOT register the user via ACME.
 func saveUser(user User) error {
 	// make user account folder
 	err := os.MkdirAll(storage.User(user.Email), 0700)
@@ -84,8 +89,10 @@ func saveUser(user User) error {
 }
 
 // newUser creates a new User for the given email address
-// with a new private key. This function does not register
-// the user via ACME.
+// with a new private key. This function does NOT save the
+// user to disk or register it via ACME. If you want to use
+// a user account that might already exist, call getUser
+// instead.
 func newUser(email string) (User, error) {
 	user := User{Email: email}
 	privateKey, err := rsa.GenerateKey(rand.Reader, rsaKeySize)
@@ -94,4 +101,47 @@ func newUser(email string) (User, error) {
 	}
 	user.key = privateKey
 	return user, nil
+}
+
+// getEmail does everything it can to obtain an email
+// address from the user to use for TLS for cfg. If it
+// cannot get an email address, it returns empty string.
+func getEmail(cfg server.Config) string {
+	// First try the tls directive from the Caddyfile
+	leEmail := cfg.TLS.LetsEncryptEmail
+	if leEmail == "" {
+		// Then try memory (command line flag or typed by user previously)
+		leEmail = DefaultEmail
+	}
+	if leEmail == "" {
+		// Then try to get most recent user email ~/.caddy/users file
+		// TODO: Probably better to open the user's json file and read the email out of there...
+		userDirs, err := ioutil.ReadDir(storage.Users())
+		if err == nil {
+			var mostRecent os.FileInfo
+			for _, dir := range userDirs {
+				if !dir.IsDir() {
+					continue
+				}
+				if mostRecent == nil || dir.ModTime().After(mostRecent.ModTime()) {
+					mostRecent = dir
+				}
+			}
+			if mostRecent != nil {
+				leEmail = mostRecent.Name()
+			}
+		}
+	}
+	if leEmail == "" {
+		// Alas, we must bother the user and ask for an email address
+		reader := bufio.NewReader(os.Stdin)
+		fmt.Print("Email address: ") // TODO: More explanation probably, and show ToS?
+		var err error
+		leEmail, err = reader.ReadString('\n')
+		if err != nil {
+			return ""
+		}
+		DefaultEmail = leEmail
+	}
+	return strings.TrimSpace(leEmail)
 }
